app/requirements.txt

fastapi
uvicorn
python-multipart
openai
load_dotenv
